Split renumber helpers and parameterize table tests
Move the footnote renumber unit tests into a dedicated submodule file
and extract the dense definition-scanning and reordering helpers into a
separate implementation module. This brings `renumber.rs` back under the
repository file-size limit without changing the top-level renumbering
flow.

Collapse the duplicated indented HTML table regression tests into a
single `rstest` so the behaviour stays covered with less repetition.
